lib: move all classifiers to `classification`
Module `classify` renamed to `classification`.
