+++ crates/rsonpath-lib/src/classification/structural/nosimd.rs
@@ -1,6 +1,8 @@
 use super::*;
+use crate::classification::quotes::{
+    QuoteClassifiedBlock, ResumeClassifierBlockState, ResumeClassifierState,
+};
 use crate::debug;
-use crate::quotes::{QuoteClassifiedBlock, ResumeClassifierBlockState, ResumeClassifierState};
 
 struct Block<'a> {
     quote_classified: QuoteClassifiedBlock<'a>,
